Update to last rustc version

+++ examples/3d/3d.rs
@@ -102,7 +102,7 @@ fn main() {
         fmod.set_3D_listener_attributes(0, &listener_pos, &vel, &forward, &up);
 
         let mut tmp = String::from_str("|.......................<1>......................<2>....................|\r");
-        unsafe { tmp.as_mut_vec().as_mut()[listener_pos.x as usize + 35usize] = 'L' as u8; }
+        unsafe { tmp.as_mut_vec().as_mut()[(listener_pos.x as isize + 35isize) as usize] = 'L' as u8; }
         print!("{}", tmp);
         fmod.update();
         ::std::thread::sleep_ms(interface_update_time as u32 - 1u32);
